Code format restructed
Dead-lock issue in thread pool testing fixed.

Removed binary file in remote repo.

Dead lock fixed(logic error).

Add cmake-build-debug to gitignore
